+++ packages/amplify-category-function/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.25.11](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-function@2.25.8...amplify-category-function@2.25.11) (2020-11-20)
+
+**Note:** Version bump only for package amplify-category-function
+
+
+
+
+
+## [2.25.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-function@2.25.8...amplify-category-function@2.25.10) (2020-11-20)
+
+**Note:** Version bump only for package amplify-category-function
+
+
+
+
+
+## [2.25.9](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-function@2.25.8...amplify-category-function@2.25.9) (2020-11-19)
+
+**Note:** Version bump only for package amplify-category-function
+
+
+
+
+
 ## [2.25.8](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-function@2.25.7...amplify-category-function@2.25.8) (2020-11-08)
 
 **Note:** Version bump only for package amplify-category-function