+++ packages/graphql-versioned-transformer/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.15.28](https://github.com/aws-amplify/amplify-cli/compare/graphql-versioned-transformer@4.15.25...graphql-versioned-transformer@4.15.28) (2020-11-20)
+
+**Note:** Version bump only for package graphql-versioned-transformer
+
+
+
+
+
+## [4.15.27](https://github.com/aws-amplify/amplify-cli/compare/graphql-versioned-transformer@4.15.25...graphql-versioned-transformer@4.15.27) (2020-11-20)
+
+**Note:** Version bump only for package graphql-versioned-transformer
+
+
+
+
+
+## [4.15.26](https://github.com/aws-amplify/amplify-cli/compare/graphql-versioned-transformer@4.15.25...graphql-versioned-transformer@4.15.26) (2020-11-19)
+
+**Note:** Version bump only for package graphql-versioned-transformer
+
+
+
+
+
 ## [4.15.25](https://github.com/aws-amplify/amplify-cli/compare/graphql-versioned-transformer@4.15.24...graphql-versioned-transformer@4.15.25) (2020-11-08)
 
 **Note:** Version bump only for package graphql-versioned-transformer